+++ app/examples/sorting/SortingExampleCustomIcons.tsx
@@ -7,7 +7,10 @@ import { useEffect, useState } from 'react';
 import { companies, type Company } from '~/data';
 
 export default function SortingExampleCustomIcons() {
-  const [sortStatus, setSortStatus] = useState<DataTableSortStatus>({ columnAccessor: 'name', direction: 'asc' });
+  const [sortStatus, setSortStatus] = useState<DataTableSortStatus<Company>>({
+    columnAccessor: 'name',
+    direction: 'asc',
+  });
   const [records, setRecords] = useState(sortBy(companies, 'name'));
 
   useEffect(() => {
@@ -24,7 +27,7 @@ export default function SortingExampleCustomIcons() {
         { accessor: 'name', width: '40%', sortable: true },
         { accessor: 'streetAddress', width: '60%' },
         { accessor: 'city', width: 160, sortable: true },
-        { accessor: 'state', sortable: true },
+        { accessor: 'state', textAlign: 'right', sortable: true },
       ]}
       sortStatus={sortStatus}
       onSortStatusChange={setSortStatus}